ref: moved to AssetRuntime
